% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airPLS.R
\name{airPLS}
\alias{airPLS}
\title{Baseline correction using airPLS (adaptive iteratively reweighted PLS).}
\usage{
airPLS(X, lambda = 1e+08, order = 2, wep = 0.1, p = 0.05, itermax = 20)
}
\arguments{
\item{X}{Row matrix of spectra/chromatograms (m rows  n samples).}

\item{lambda}{Smoothness penalty. Default 1e8.}

\item{order}{Difference penalty order. Default 2.}

\item{wep}{Weight exception proportion at endpoints. Default 0.1.}

\item{p}{Asymmetry parameter at endpoints. Default 0.05.}

\item{itermax}{Maximum iterations. Default 20.}
}
\value{
List: \code{$Xc} corrected matrix, \code{$Z} baseline matrix.
}
\description{
Baseline correction using airPLS (adaptive iteratively reweighted PLS).
}
