% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads.R
\name{beads}
\alias{beads}
\title{BEADS baseline estimation.}
\usage{
beads(
  y,
  d = 1,
  fc = 0.01,
  r = 6,
  lam0 = 0.5,
  lam1 = 5,
  lam2 = 4,
  pen = "L1_v2",
  Nit = 30L
)
}
\arguments{
\item{y}{Numeric vector  raw chromatogram signal.}

\item{d}{Filter order (1 or 2).}

\item{fc}{Filter cut-off frequency (cycles/sample, 0 < fc < 0.5).}

\item{r}{Asymmetry ratio.}

\item{lam0}{Regularisation parameter (sparsity of x).}

\item{lam1}{Regularisation parameter (sparsity of first derivative).}

\item{lam2}{Regularisation parameter (sparsity of second derivative).}

\item{pen}{Penalty: "L1_v1" (smooth L1) or "L1_v2" (log-barrier L1).}

\item{Nit}{Number of iterations (default 30).}
}
\value{
List: $x (denoised signal), $f (baseline), $cost (cost history).
}
\description{
BEADS baseline estimation.
}
