% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibit.R
\name{inhibit}
\alias{inhibit}
\title{Define inhibit regions where baseline fitting is suppressed.}
\usage{
inhibit(RT, inhibit_regions = list())
}
\arguments{
\item{RT}{Numeric retention time vector (minutes).}

\item{inhibit_regions}{List of \code{c(t_start, t_end)} pairs (minutes).}
}
\value{
List: \code{$block} logical mask, \code{$segments} list of good-region index pairs.
}
\description{
Define inhibit regions where baseline fitting is suppressed.
}
