% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{meta}
\alias{meta}
\title{Extract run metadata from a ChromResult.}
\usage{
meta(object)
}
\arguments{
\item{object}{A \code{ChromResult} object.}
}
\value{
A named list of run metadata and method settings.
}
\description{
Extract run metadata from a ChromResult.
}
