% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{peaks}
\alias{peaks}
\title{Extract the peaks table from a ChromResult.}
\usage{
peaks(object)
}
\arguments{
\item{object}{A \code{ChromResult} object.}
}
\value{
A data.frame of detected peaks.
}
\description{
Extract the peaks table from a ChromResult.
}
