% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{traces}
\alias{traces}
\title{Extract per-chromatogram signal/baseline traces from a ChromResult.}
\usage{
traces(object)
}
\arguments{
\item{object}{A \code{ChromResult} object.}
}
\value{
A list of trace data.frames (one per chromatogram).
}
\description{
Extract per-chromatogram signal/baseline traces from a ChromResult.
}
